AX: [ATK] ARIA menu items should not have anonymous block children https://bugs.webkit.org/show_bug.cgi?id=144653
Reviewed by Chris Fleizach.

Source/WebCore:

Exclude anonymous block children of menu items from the accessible tree
for ATK. Exclude list markers from the accessibility tree if the parent
is not being exposed as an accessible list item.

No new tests. Instead, modified the existing tests to reflect the correct
behavior.

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::computeAccessibilityIsIgnored):
* accessibility/atk/AccessibilityObjectAtk.cpp:
(WebCore::AccessibilityObject::accessibilityPlatformIncludesObject):

LayoutTests:

The existing test was not cross-platform friendly and was rewritten as
part of the fix for this bug. New expectations were generated for Mac,
Efl, and Gtk. The shared expectations and the Windows expectations were
removed so as not to break the Win port. The search-predicate test also
was updated because it included an ARIA treeitem which had a listmarker
child it should not have, and no longer has as a result of the bug fix.

* accessibility/aria-menubar-menuitems-expected.txt: Removed. Platform hierarchies vary vastly.
* accessibility/aria-menubar-menuitems.html: Rewritten to work better as a shared test.
* platform/efl/accessibility/aria-menubar-menuitems-expected.txt: Added.
* platform/gtk/accessibility/aria-menubar-menuitems-expected.txt: Added.
* platform/mac/accessibility/aria-menubar-menuitems-expected.txt: Modified as result of rewrite.
* platform/mac/accessibility/search-predicate.html: Modified to reflect correct treeitem children.
* platform/mac/accessibility/search-predicate-expected.txt: Modified to reflect correct treeitem children.
* platform/win/accessibility/aria-menubar-menuitems-expected.txt: Removed. New file should be generated.
* resources/accessibility-helper.js: Added an argument so dumpAccessibilityTree can print title rather than value.
(dumpAccessibilityTree):


Identifier: 162964@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@184213 268f45cc-cd09-0410-ab3c-d52691b4dbfc