REGRESSION (r181910): WKWebView incorrectly scales snapshot https://bugs.webkit.org/show_bug.cgi?id=145076
Patch by James Savage <james.savage@apple.com> on 2015-05-15
Reviewed by Tim Horton.

We added a fast path to snapshotting using IOSurfaces with
http://trac.webkit.org/changeset/181910 which incorrectly determined
scale and transform and resulted in bad snapshots in some situations.

* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _snapshotRect:intoImageOfWidth:completionHandler:]):
When snapshotting an IOSurface we need to compute our scale based off of
the rect in the WKWebView coordinates, we also failed to account for
non-zero origins in the snapshot rect. Remove an unused varible while
we're here.

Identifier: 163110@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@184422 268f45cc-cd09-0410-ab3c-d52691b4dbfc