Don't reset the preview recognizer in [WKContentViewInteraction cleanupInteraction] https://bugs.webkit.org/show_bug.cgi?id=145081
We shouldn't reset the recognizer in [WKContentViewInteraction cleanupInteraction] since we don't re-add it
back to WKContentView when the process relaunched. Since we already add/remove it when the view is move into/
removed from the window, we don't need reset it in cleanupInteraction.

Patch by Yongjun Zhang <yongjun_zhang@apple.com> on 2015-05-15
Reviewed by Dan Bernstein.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView cleanupInteraction]):

Identifier: 163113@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@184425 268f45cc-cd09-0410-ab3c-d52691b4dbfc