AX: improve list heuristics (presentational use versus actual lists) https://bugs.webkit.org/show_bug.cgi?id=134187
Unreviewed layout test fixage.

Skip the list-detection2.html test on GTK until it can be investigated.
    https://bugs.webkit.org/show_bug.cgi?id=145273
Fix a failing test on GTK that needed an update after the last set of changes.
Skip the test on windows until a time someone can decide if we even expose or need lists in Windows AX API.
    https://bugs.webkit.org/show_bug.cgi?id=145271

* platform/gtk/TestExpectations:
* platform/gtk/accessibility/list-detection-expected.txt:
* platform/win/TestExpectations:



Identifier: 163283@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@184721 268f45cc-cd09-0410-ab3c-d52691b4dbfc