Web Inspector: Uncaught exception when using Inspect tool on SVG elements https://bugs.webkit.org/show_bug.cgi?id=145363
Reviewed by Joseph Pecoraro.

The injected script failed by chaining a call to String.prototype.trim to the result of
SVG*Element.className, which is an SVGAnimatedString and lacks useful methods. So, obtain
the class name using Node.getAttribute, which always returns a DOMString.

* inspector/InjectedScriptSource.js:
(InjectedScriptSource.prototype._getDescription): use getAttribute instead of className.

Identifier: 163363@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@184847 268f45cc-cd09-0410-ab3c-d52691b4dbfc