GraphicsLayerCA should start out with m_intersectsCoverageRect=false <https://webkit.org/b/145443> <rdar://problem/19685798>
Reviewed by Simon Fraser.

Source/WebCore:

Since the coverage rect of a newly-constructed GraphicsLayerCA is always empty,
there's no way that the layer's own bounding box can intersect with it.
Hence the m_intersectsCoverageRect flag should be false after construction.

This fixes an issue where compositing layers in subframes would be marked as
non-volatile, despite being far outside the viewport and coverage rect.

Test: compositing/visible-rect/iframe-with-layers-outside-viewport.html

* platform/graphics/ca/GraphicsLayerCA.cpp:
(WebCore::GraphicsLayerCA::GraphicsLayerCA):

LayoutTests:

Add a test (based on iframe-and-layers.html) that makes an iframe with compositing
layers inside it and places it far outside the visible viewport.

The layer dump should show that the "intersects coverage rect" flag is 0 for the
iframe and its sublayers.

* compositing/visible-rect/iframe-with-layers-outside-viewport-expected.txt: Added.
* compositing/visible-rect/iframe-with-layers-outside-viewport.html: Added.

Identifier: 163472@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@184968 268f45cc-cd09-0410-ab3c-d52691b4dbfc